K3
  * Changed BW selector to include finer steps
  * Added read options to initialization method.
    - set power control iaw availability of K3/100 module
    - required a change in get_pc_min_max method - forced
      change in non K3 rig classes
  * Added default Mode/BW table
  * Changed Power out method to use the BG; query

+++ src/include/K2.h
@@ -22,7 +22,7 @@ public:
 	int get_bwA();
 	void tune_rig();
 	void set_PTT_control(int val);
-    virtual void get_pc_min_max_step(int &min, int &max, int &step){
+    virtual void get_pc_min_max_step(double &min, double &max, double &step){
         min = 0; max = 110; step = 1; }
 	int  get_power_control();
 	void set_power_control(double val);
